chore: handle chat functions
