+++ web-app/src/hooks/useAppState.ts
@@ -1,20 +1,27 @@
 import { create } from 'zustand'
 import { ThreadMessage } from '@janhq/core'
+import { MCPTool } from '@/types/completion'
 
 type AppState = {
   streamingContent?: ThreadMessage
   loadingModel?: boolean
+  tools: MCPTool[]
   updateStreamingContent: (content: ThreadMessage | undefined) => void
   updateLoadingModel: (loading: boolean) => void
+  updateTools: (tools: MCPTool[]) => void
 }
 
 export const useAppState = create<AppState>()((set) => ({
   streamingContent: undefined,
   loadingModel: false,
+  tools: [],
   updateStreamingContent: (content) => {
     set({ streamingContent: content })
   },
   updateLoadingModel: (loading) => {
     set({ loadingModel: loading })
+  },
+  updateTools: (tools) => {
+    set({ tools })
   },
 }))
